+++ src/styles/raw-themes/light-raw-theme.js
@@ -14,7 +14,7 @@ module.exports = {
   palette: {
     primary1Color: Colors.cyan500,
     primary2Color: Colors.cyan700,
-    primary3Color: Colors.lightBlack,
+    primary3Color: Colors.grey400,
     accent1Color: Colors.pinkA200,
     accent2Color: Colors.grey100,
     accent3Color: Colors.grey500,