+++ src/styles/transitions.js
@@ -32,7 +32,7 @@ module.exports = {
     delay = delay || '0ms';
     easeFunction = easeFunction || "linear";
 
-    return AutoPrefix.singleHyphened(property) + ' ' +
+    return property + ' ' +
       duration + ' ' +
       easeFunction + ' ' +
       delay;